Major code overhaul
Insertion of global variables to optimize speed and stabilize connection with Wikidata.

Question Selector for F01 is good

answer insert for F01 is WIP

agent insert function is good.

individual answer and cluster answer is WIP.

use of select2 library for enabling search in dropdown lists.

use of shortuuid library for creating uuids in the RLWF triple-store.

updates and optimizations in all local and WDT queries.

inserted a few prints to make sure if the code is going a specific way.

I think this is the biggest update so far. there is still a lot to do but we are almost there.

Updates in the python notebooks as well as the ontology(as always)